requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
jinja2==3.1.6
python-slugify==8.0.4
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.2.0
pydantic-settings==2.10.1
